+++ pages/api/user/index.ts
@@ -4,26 +4,31 @@ import type { NextApiRequest, NextApiResponse } from "next";
 const prisma = new PrismaClient();
 
 export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse
+	req: NextApiRequest,
+	res: NextApiResponse
 ) {
-  if (req.method === "GET") {
-    const id = req.query.identifier as string;
-    try {
-      let user: User | null;
-      if (id) {
-        user = await prisma.user.findUnique({
-          where: {
-            identifier: id,
-          },
-        });
-      } else {
-        throw new Error("user not found");
-      }
-      const data = res.status(200).json(user);
-      return data;
-    } catch (err) {
-      console.log(err);
-    }
-  }
+	if (req.method === "GET") {
+		const id = req.query.identifier as string;
+		try {
+			let user: User | null = null;
+			if (id) {
+				user = await prisma.user.findUnique({
+					where: {
+						identifier: id,
+					},
+				});
+			} else {
+				res.status(500).send("User ID is required");
+			}
+
+			if (user) {
+				res.status(200).json(user);
+			} else {
+				res.status(404).send("user not found");
+			}
+		} catch (err) {
+			console.log(err);
+			res.status(500).end();
+		}
+	}
 }
